+++ SQLXEtoEventHub/SQLXEtoEventHub.csproj
@@ -54,10 +54,10 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="EventHubWriter.cs" />
-    <Compile Include="PositionRegistryPersist.cs" />
+    <Compile Include="XEPosition\RegistryStore.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="XEPosition.cs" />
+    <Compile Include="XEPosition\XEPosition.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
