+++ SQLXEtoEventHub/XEPosition/XEPosition.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SQLXEtoEventHub
+namespace SQLXEtoEventHub.XEPosition
 {
     public class XEPosition
     {
