+++ Test/Test.csproj
@@ -52,7 +52,7 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
-    <Compile Include="PositionRegistryPersist_Test.cs" />
+    <Compile Include="RegistryStore_Test.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
